+++ src/algorithms/column_generation.h
@@ -23,7 +23,7 @@ struct CGResult {
     std::vector<double> solution_costs;
     std::vector<double> integer_solution_costs;
     std::vector<double> covered_interventions;
-    std::vector<int> integer_covered_interventions;
+    std::vector<double> integer_covered_interventions;
     std::vector<double> time_points;
 };
