+++ src/data_analysis/plot.cpp
@@ -106,8 +106,8 @@ void plot_objective_values(
 
 void plot_objective_values(
     matplot::axes_handle& ax,
-    const std::vector<int>& time_points,
-    const std::vector<int>& values,
+    const std::vector<double>& time_points,
+    const std::vector<double>& values,
     const std::string& name,
     bool use_semilogy,
     int N
